Exercises 4.21.

+++ part4/blog_list/controllers/blogs.js
@@ -33,7 +33,21 @@ blogRouter.post("/", async (request, response) => {
 });
 
 blogRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
   await Blog.findByIdAndDelete(request.params.id);
+
+  user.blogs = user.blogs.filter(
+    (blogId) => blogId.toString() !== request.params.id
+  );
+
+  console.log("updatedUser", user);
+  await user.save();
+
   response.status(204).end();
 });
